use tarfiles instead of xattr to bind datalake files and metadata
Why? symlinks aren't going to work. Consider the case in which
logrotate is enqueuing logs to the queue directory. If the
uploader crashes or otherwise fails to process one of the
enqueued files, the next logrotate run will overwrite the failed
file. Not only will this prevent recovery, it will cause
confusion. The disk overhead of this scheme is a noted cost that
we accept.