expose s3 events to ingester
Why? Because we are about to (temporarily?) support moving files
in the s3 bucket so we can do some rearranging of the s3
urls. For this, we'll need to differentiate between S3's Put and
Copy. This inspired some refactoring. Specifically, we no longer
need a "translator". Instead, we have an S3Notification that
contains some number of "Records" (which are always "S3Events"
when we care about them). This matches s3's model better anyway.