use moto instead of dynamodb_local
We're about to use moto's s3 mocking. So we add moto. And
httpretty, which is a dependency of moto, breaks
dynamodb_local. (In fact, it generally seems to break lots of
non-mocked network things). So we just depend on moto for
everything. This also eliminates the need for vagrant, which is
nice.

Note that there is a minor behavior change. dynamodb_local was
not emulating global secondary indexes properly. Specifically, if
multiple documents map to the same GSI hash and range key,
dynamodb_local only returns one of them but should return
all. See
http://docs.aws.amazon.com/amazondynamodb/latest/developerguide/GSI.html. moto's
behvior is correct.