add queries for latest files
Users often need to retrieve the latest file for a given where
and what. This is a quick solution that hides what these users
are currently doing behind the API. It's not perfect, of
course. It only looks back 14 days for the latest file. But it's
a start. And if this turns out to be insufficient, we can add
something bigger and better (e.g., a cache of recent files,
another dynamodb index, etc.)