do not rely on filename to detect content encoding
Instead, inspect the header, which is more definitive. Note that
I considered using python-magic to achieve this, which may have
left the door open for supporting additional content types and
encodings in the future. But on some platforms the tests would
fail despite a successful install of python-magic. Perhaps this
has something to do with the underlying libmagic configuration.
But I don't want to fight that wildcard behavior. So I baked in
the specific thing that we need.