support python3
Much of this patch is just applying some basic conventions (e.g.,
using six, using relative imports, etc.). But file and string IO
require a bit more nuance. Specifically, we must be explicit
about character encoding. To this end, I followed the rule that
when writing or reading from files or web sockets, always
expect/use bytes (and BytesIO); do not assume strings (or
StringIO). I also added a test to verify that a legacy bundle
(which is a data file format entirely internal to the datalake
client) can still be handled.