compose dynamodb queries more explicitly
On rare occasions, the boto3.dynamodb.conditions generate invalid
conditions. Here is an example query payload from the field:

    {
        "KeyConditionExpression": "(#n0 = :v0 AND begins_with(#n0, :v0))",
        "TableName": "the-table",
        "Limit": 100,
        "ExpressionAttributeValues": {":v0": {"S": "somewhere:"}},
        "ExpressionAttributeNames": {"#n0": "range_key"}
    }

This is from a query of the form:

/v0/archive/files?what=somewhat&where=somewhere&start=123&end=456

The `#n0 = :v0` business in the query is supposed to be for the
time bucket. But as you can see from the ExpressionAttribute*
portions of the query, #n0 and :v0 have collided with the range
key business in the query. This results in a `ClientError` with
the message:

An error occurred (ValidationException) when calling the Query
operation: KeyConditionExpressions must only contain one
condition per key.

I was unable to reproduce these conditions in a dev env, and was
not able to find the root cause otherwise. Instead, I make the
code more explicit and hopefully deterministic in how it prepares
the queries. If this doesn't fix the problem, it will at least
improve visibility.