drop boto2 from client
Note that this work relies heavily on work by @lod in
https://github.com/planetlabs/datalake/pull/78/

Note that the ingester and api both need this treatment too. And
both the ingester and the api rely on the datalake client for
shared code (e.g., Record, Metadata, various test fixtures). But
migrating just the client was a heavy enough lift. To preserve
the ingester and api funtionality, I made the docker container
install datalake<1. We will migrate these back to the main line
code in subsequent PRs.